Updated string parsing function to enforce stricter rules for UTF-8
Also fixed issue where the column counter was advanced for each byte
in a UTF-8 sequence, now the counter only increments for an entire
character, regardless of encoding width. Note that this doesn't
apply for UTF-16 escape sequences in UTF-8 strings, in that case
each character of the escape sequence advances the column.

Also, previously I was permitting ASCII-del (0x7f) characters in
strings, which is the only ASCII control character above 0x1f.